Added versioning Information

+++ BIDebugEngine/BIDebugEngine/EngineHook.h
@@ -22,7 +22,7 @@ public:
 
     struct Pattern {
         Pattern(const char* _mask, const char* _pattern) : mask(_mask), pattern(_pattern) {}
-        Pattern(const char* _mask, const char* _pattern, int64_t _offset) : mask(_mask),pattern(_pattern),offset(_offset) {}
+        Pattern(const char* _mask, const char* _pattern, int64_t _offset) : mask(_mask),pattern(_pattern),offset(static_cast<uintptr_t>(_offset)) {}
         const char* mask;
         const char* pattern;
         uintptr_t offset{ 0 };
